+++ src/LiveComponent/src/Twig/LiveComponentExtension.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\UX\LiveComponent\Twig;
 
+use Twig\DeprecatedCallableInfo;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
@@ -24,8 +25,19 @@ final class LiveComponentExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('component_url', [LiveComponentRuntime::class, 'getComponentUrl']),
-            new TwigFunction('live_action', [LiveComponentRuntime::class, 'liveAction'], ['is_safe' => ['html_attr']]),
+            new TwigFunction('component_url', [LiveComponentRuntime::class, 'getComponentUrl'], [
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-live-component', '3.1', 'ux_live_component_url')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-live-component', 'alternative' => 'ux_live_component_url']),
+            ]),
+            new TwigFunction('ux_live_component_url', [LiveComponentRuntime::class, 'getComponentUrl']),
+            new TwigFunction('live_action', [LiveComponentRuntime::class, 'liveAction'], [
+                'is_safe' => ['html_attr'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-live-component', '3.1', 'ux_live_action')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-live-component', 'alternative' => 'ux_live_action']),
+            ]),
+            new TwigFunction('ux_live_action', [LiveComponentRuntime::class, 'liveAction'], ['is_safe' => ['html_attr']]),
         ];
     }
 }