+++ src/Turbo/src/Twig/TwigExtension.php
@@ -29,22 +29,75 @@ final class TwigExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('turbo_is_frame_request', [TurboRuntime::class, 'isTurboFrameRequest']),
-            new TwigFunction('turbo_frame_request_id', [TurboRuntime::class, 'getTurboFrameRequestId']),
-            new TwigFunction('turbo_stream_from', [TurboRuntime::class, 'renderTurboStreamFrom'], ['needs_environment' => true, 'is_safe' => ['html']]),
+            new TwigFunction('turbo_is_frame_request', [TurboRuntime::class, 'isTurboFrameRequest'], [
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_is_frame_request')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_is_frame_request']),
+            ]),
+            new TwigFunction('ux_turbo_is_frame_request', [TurboRuntime::class, 'isTurboFrameRequest']),
+            new TwigFunction('turbo_frame_request_id', [TurboRuntime::class, 'getTurboFrameRequestId'], [
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_frame_request_id')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_frame_request_id']),
+            ]),
+            new TwigFunction('ux_turbo_frame_request_id', [TurboRuntime::class, 'getTurboFrameRequestId']),
+            new TwigFunction('turbo_stream_from', [TurboRuntime::class, 'renderTurboStreamFrom'], [
+                'needs_environment' => true,
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_stream_from')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_stream_from']),
+            ]),
+            new TwigFunction('ux_turbo_stream_from', [TurboRuntime::class, 'renderTurboStreamFrom'], ['needs_environment' => true, 'is_safe' => ['html']]),
             new TwigFunction('turbo_stream_listen', [TurboRuntime::class, 'renderTurboStreamListen'], [
                 'needs_environment' => true,
                 'is_safe' => ['html'],
                 ...(class_exists(DeprecatedCallableInfo::class)
-                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'turbo_stream_from')]
-                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'turbo_stream_from']),
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_stream_from')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_stream_from']),
             ]),
-            new TwigFunction('turbo_exempts_page_from_cache', $this->turboExemptsPageFromCache(...), ['is_safe' => ['html']]),
-            new TwigFunction('turbo_exempts_page_from_preview', $this->turboExemptsPageFromPreview(...), ['is_safe' => ['html']]),
-            new TwigFunction('turbo_page_requires_reload', $this->turboPageRequiresReload(...), ['is_safe' => ['html']]),
-            new TwigFunction('turbo_refreshes_with', $this->turboRefreshesWith(...), ['is_safe' => ['html']]),
-            new TwigFunction('turbo_refresh_method', $this->turboRefreshMethod(...), ['is_safe' => ['html']]),
-            new TwigFunction('turbo_refresh_scroll', $this->turboRefreshScroll(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_exempts_page_from_cache', $this->turboExemptsPageFromCache(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_exempts_page_from_cache')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_exempts_page_from_cache']),
+            ]),
+            new TwigFunction('ux_turbo_exempts_page_from_cache', $this->turboExemptsPageFromCache(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_exempts_page_from_preview', $this->turboExemptsPageFromPreview(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_exempts_page_from_preview')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_exempts_page_from_preview']),
+            ]),
+            new TwigFunction('ux_turbo_exempts_page_from_preview', $this->turboExemptsPageFromPreview(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_page_requires_reload', $this->turboPageRequiresReload(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_page_requires_reload')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_page_requires_reload']),
+            ]),
+            new TwigFunction('ux_turbo_page_requires_reload', $this->turboPageRequiresReload(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_refreshes_with', $this->turboRefreshesWith(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_refreshes_with')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_refreshes_with']),
+            ]),
+            new TwigFunction('ux_turbo_refreshes_with', $this->turboRefreshesWith(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_refresh_method', $this->turboRefreshMethod(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_refresh_method')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_refresh_method']),
+            ]),
+            new TwigFunction('ux_turbo_refresh_method', $this->turboRefreshMethod(...), ['is_safe' => ['html']]),
+            new TwigFunction('turbo_refresh_scroll', $this->turboRefreshScroll(...), [
+                'is_safe' => ['html'],
+                ...(class_exists(DeprecatedCallableInfo::class)
+                    ? ['deprecation_info' => new DeprecatedCallableInfo('symfony/ux-turbo', '3.1', 'ux_turbo_refresh_scroll')]
+                    : ['deprecated' => '3.1', 'deprecating_package' => 'symfony/ux-turbo', 'alternative' => 'ux_turbo_refresh_scroll']),
+            ]),
+            new TwigFunction('ux_turbo_refresh_scroll', $this->turboRefreshScroll(...), ['is_safe' => ['html']]),
         ];
     }